Feature/Fix: 비밀번호 재설정하기, 변경하기 및 계정 삭제 로직 변경

+++ YeDi/YeDi/Shared/ViewModel/Auth/AuthViewModel.swift
@@ -29,7 +29,7 @@ final class UserAuth: ObservableObject {
     }
     
     func fetchUser() {
-        auth.addStateDidChangeListener { auth, user in
+        auth.addStateDidChangeListener { _, user in
             if let user = user {
                 self.userSession = user
                 self.userType = self.userType
@@ -68,7 +68,6 @@ final class UserAuth: ObservableObject {
             if let error = error {
                 print("DEBUG: signIn Error \(error.localizedDescription)")
                 completion(false)
-                return
             }
             
             guard let user = result?.user else {
@@ -183,6 +182,37 @@ final class UserAuth: ObservableObject {
         }
     }
     
+    func resetPassword(forEmail email: String, completion: @escaping (Bool) -> Void) {
+        auth.sendPasswordReset(withEmail: email) { error in
+            if let error = error {
+                print(error.localizedDescription)
+                completion(false)
+            } else {
+                completion(true)
+            }
+        }
+    }
+    
+    func updatePassword(_ email: String, _ currentPassword: String, _ newPassword: String, _ completion: @escaping (Bool) -> Void) {
+        let credential: AuthCredential = EmailAuthProvider.credential(withEmail: email, password: currentPassword)
+        
+        auth.currentUser?.reauthenticate(with: credential) { result, error in
+            if let error = error {
+                print("reauthenticate error: \(error.localizedDescription)")
+                completion(false)
+            } else {
+                self.auth.currentUser?.updatePassword(to: newPassword) { error in
+                    if let error = error {
+                        print("update error: \(error.localizedDescription)")
+                        completion(false)
+                    } else {
+                        completion(true)
+                    }
+                }
+            }
+        }
+    }
+    
     func signOut() {
         userSession = nil
         userType = nil
@@ -195,12 +225,29 @@ final class UserAuth: ObservableObject {
     }
     
     func deleteClientAccount() {
+        guard let user = Auth.auth().currentUser else { return }
+        
         if let currentClientID {
             storeService
                 .collection("clients")
                 .document(currentClientID).delete()
+        }
+        
+        if let currentDesignerID {
+            storeService
+                .collection("designers")
+                .document(currentDesignerID).delete()
+        }
+        
+        user.delete { error in
+            if let error = error {
+                print("DEBUG: Error deleting user account: \(error.localizedDescription)")
+                return
+            }
             
-            userSession = nil
+            print("DEBUG: User account deleted")
+            
+            self.signOut()
         }
     }
 }
